requirements.txt
discord.py==2.4.0
aiohttp==3.9.1
flask==3.0.0
requests==2.31.0
certifi==2024.2.2
PyNaCl==1.5.0
nest_asyncio==1.6.0
audioop-lts==0.2.1
Pillow==10.1.0
easy-pil==0.3.3
